ai_parser_service/requirements.txt
# Core Dependencies
openai==1.35.13
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0

# Database
sqlalchemy==2.0.21

# Utilities
requests==2.31.0
